Fixed JS My intial js file logged all e.target.innerText for every button into my input Array which forced me to split the array with regex to evaluate the inputs While this worked for basic operations it failed upon introducing negative values as a null value would appear for every charcter that was split, failing for this meant programming other function buttons to be unfeasible. Upon realizing this, I started a new js file that seperated numbers and operators into their respective arrays. This allowed for functions to be seperated into more concise methods.
logic2 finished all basic operations, +/-, and  % buttons.
Working on functionality of the decimal button, deciding whether
to search array based on a regex for'.' or to simply disable
the button upon click.
To start:
- A backspace button that allows user to delete their input,
which is the reason I am deliberating upon the decimal function as it
will have to be enabled if deleted in the user input.

- round answers to prevent overflow, I have delayed this as I have
yet to start css layout thus unsure as to the limit of digits in the
display

- css format and design

- keyboard support

- divide by 0